🎨 [PANA-6776] Separate session replay test helpers and update CODEOWNERS (#4435)
