+++ packages/rum/src/domain/record/serialization/serializeAttributes.spec.ts
@@ -17,7 +17,7 @@ import type { VirtualAttributes } from './serialization.types'
 import type { SerializationMetric, SerializationStats } from './serializationStats'
 import { createSerializationStats } from './serializationStats'
 
-const CSS_FILE_URL = '/base/packages/rum/test/toto.css'
+const CSS_FILE_URL = '/base/packages/rum/test/record/toto.css'
 
 const PRIVACY_LEVELS = Object.keys({
   [NodePrivacyLevel.ALLOW]: true,
