+++ packages/rum/test/index.ts
@@ -1,8 +1,4 @@
 export * from './mockWorker'
-export * from './recordsPerFullSnapshot'
-export * from './mutationPayloadValidator'
-export * from './nodes'
-export * from './segments'
-export * from './readReplayPayload'
+export * from './record'
 export * from './rumFrustrationEvent'
 export * from './mockProfiler'
