+++ packages/rum/test/record/index.ts
@@ -1,0 +1,5 @@
+export * from './recordsPerFullSnapshot'
+export * from './mutationPayloadValidator'
+export * from './nodes'
+export * from './segments'
+export * from './readReplayPayload'
