+++ packages/rum/test/record/readReplayPayload.ts
@@ -1,7 +1,7 @@
 import type { Payload } from '@datadog/browser-core'
-import type { BrowserSegment } from '../src/types'
-import type { BrowserSegmentMetadataAndSegmentSizes } from '../src/domain/segmentCollection'
-import { readFormData } from '../../core/test'
+import type { BrowserSegment } from '../../src/types'
+import type { BrowserSegmentMetadataAndSegmentSizes } from '../../src/domain/segmentCollection'
+import { readFormData } from '../../../core/test'
 
 export function readReplayPayload(payload: Payload) {
   return readFormData<{
