+++ packages/rum/test/record/recordsPerFullSnapshot.ts
